Added weapon ray tracing and hit detection
A helper function called WeaponTracing will do a simple ray trace from the controller and detect for any hits. It is also optionally pressure-sensitive, though that may not be necessary until different weapons are implemented. Also set it up so that the debug lines and spheres can be toggled in-editor and don't require a recompile to hide.