Renaming an important function
We had a function called Respawn Doors which would reset the progression and start back from Door 0 in terms of location, difficulty, etc. This was called "[Editor] Respawn Doors" because it was originally intended to be a button you just clicked in the editor. However, it is so complete that we can call this when the player dies and it will flawlessly regenerate the doors, any time.