Adding collision + Floor Offset to mesh
If the hall meshes origin is at the very base of the mesh, then we'll have to offset both the doors and the platform so that they rest on top of the tunnel's floor (which is raised up). I changed it so that the floor of the tunnel is where the origin lies. Also added sweet new debug features because I like seeing all those cool neon wireframes.