Crystal Color now represents total HP
Instead of brightness representing current HP percent and Color representing initial HP, the color simply represents HP on some static scale (such that 0 HP is green, half HP on the scale is red, and full HP on the scale is blue.) That maximum value for the color scale is exposed as a variable (Color Maximum Health). I added a python script demonstrating the math. Brightness is now also easier to configure.