Tighten up a fully functional leaderboard (!!!)
We now have a leaderboard where players can see at the end of the game how they did and leave their personalized mark in history. The leaderboard itself is of arbitrary length and the  user can scroll through it - though we do still have a way of defining the top N scores as being high scores.

Now, to test it in VR...